refactor: Update snort.conf HOME_NET configuration in test_install.py
The code changes in `test_install.py` ensure that the `HOME_NET` variable in `snort.conf` is correctly configured. The previous implementation used the `ip` command with the `-o` and `-f inet` options to retrieve the IP address, but it didn't work as expected. The updated code now uses the `ip` command with the `-4` option and `grep` to extract the IP address. The `HOME_NET` variable is then updated in `snort.conf` to include the IP address with a `/24` subnet mask.