refactor: Update snort.ps1 to use UTF-8 encoding when writing rules to file
The `Install-Snort` function in `snort.ps1` is updated to use UTF-8 encoding when writing the rules to the file. Previously, the rules were written using the default encoding, which could cause issues with certain characters. This change ensures that the rules file is properly encoded and can handle a wider range of characters.

Refactor `Install-Snort` function in `snort.ps1` to use UTF-8 encoding when writing rules to file