refactor: Update snort.ps1 to replace snort.conf file
The `snort.ps1` script is updated to replace the existing `snort.conf` file with a new version downloaded from a remote repository. This change ensures that the latest configuration file is used for Snort. If the download fails, an appropriate error message is displayed.

Refactor `snort.ps1` to replace snort.conf file