refactor: Update snort.conf output settings
The `snort.conf` file is modified to include the following output settings:
- `output alert_syslog: LOG_AUTH LOG_ALERT`
- `output alert_fast: snort.alert`

These settings enable Snort to send alerts to the syslog with the `LOG_AUTH` facility and `LOG_ALERT` priority, as well as output alerts to the `snort.alert` file. This change enhances the logging and alerting capabilities of Snort.

Refactor `snort.conf` to update output settings