refactor: Update snort.ps1 to update URLs for local.rules and snort.conf
The `snort.ps1` script is updated to replace the URLs for the `local.rules` and `snort.conf` files with the correct ones from the ADORSYS-GIS/wazuh-snort repository. This change ensures that the latest versions of these files are used for Snort. The previous URLs are no longer valid and needed to be updated.

Refactor `snort.ps1` to update URLs for local.rules and snort.conf